test(ci): keep Linux routing coverage in workflow guards
